feat/DSTEW-2081 Consolidated ReadModel update so reused by new approach and existing projection. Added a QueryHandler instead of just calling the service to follow CQRS pattern
